Add region search function (v0.3)
Added site and region search.
Site search finds the site gameobject, adds a target reticle on it, and centers the camera on it.
Region function finds the average center of all vertices in a site region, duplicates the mesh, adds a red material and makes the gameobject flash, it also tries to center on the average center of the mesh although that might not be within the actual mesh.